Load config.js after onload
We turn logn_config.js into a full plugin so that we can attach to the
page_ready event and load config.js after that fires.  This makes sure
that page load does not block waiting for config.js to load up.  In most
cases config.js loads up fast, but there may be cases where TCP connect
takes a long time and page load blocks on this.

Note that we cannot send the beacon until config.js has loaded because
it contains the crumb.  It will also call init() a second time with new
configs.