Split bandwidth and rt plugins into separate files
The core plugins no longer need to be in the boomerang.js file.  We hope
that at this point most people using boomerang have a build process
where they combine and minify their JavaScript, and this should be done
at that step.

The Makefile will by default include rt.js and bw.js, but this can be
overridden on the command line while running make.