Reference BOOMR.window instead of w where possible
In most cases where we use the window object, we use it inside a method
that's called through or after init().  Since init() is always called
from boomerang.js, we know for certain that BOOMR.window has been
defined at this point, so we can use BOOMR.window instead of w in all
these cases.