Adapt if w is not the right window
w and d are used in several locations across the rt.js plugin, and they
are defined in global space.  This makes it hard to change the reference
in one place.  The only place we can change it is in the init() call,
since this is always the first method called.

We check if the passed in w object matches the predefined BOOMR.window
object, and if not, redefine it so we have the right reference.