Add config.js to same frame as boomerang.js
config.js should be added to the same frame/window that boomerang.js is
in.  This is so that config.js does not block onload of the parent
frame.

BOOMR.window will reference the parent frame in cases where boomerang is
loaded into an iframe, so we can't use BOOMR.window.document, and have
to use document instead.