Load config.js immediately if running in iframe
If we're running inside an iframe, then we don't care about blocking
onload (since the iframe's onload even has already fired).  We can load
config.js as soon as we're ready to load it.

This also means that config.js might load before onload in the parent
window fires, which gives us the opportunity to capture unloads that
happen before onload.

I like this.