Prevent empty first argument
The var loop prepends an & to every k-v pair, and appending this to a
url that already ends with a ? or a & results in a url that looks like
this:

http://foo.com/?&v=...

or

http://foo.com/?something=blah&&v=...

ie, we get a ?& or a && in the URL, which results in an empty parameter.

This patch rewrites the url generating code to use an array and then
array.join to make sure the query string part does not have an & prefix.