Approximate fb time from on unload/pagehide
First byte time is about 1ms off from when the unload or pagehide
event fires.  This is so because browsers do not unload a page from
the browser window until they have something to display from the new
page.  This gives us a good approximation of the first byte time
since that's when browser have something to start doing.

We set the unload time and the click time in the cookie, and use the
best one as the load start time.  unload is preferred, click is used
if it's unavailable (Opera only)