Get data from cookie when getting from navtiming
It turns out that in-page JavaScript does not execute on Safari
when using the Back-Forward buttons.  This has to do with the page
state being stored in the PageCache.  To get around this, we need
to recheck the cookie every time we need a start time.

We also need to reset t_start and t_fb_approx to undefined if the
cookie's values were inconclusive.  It's important that the order
of calls to initFromCookie and initNavTiming be such since
initNavTiming will always have the correct time if navtiming is
available.

Conflicts:
	plugins/rt.js