Make sessionID configurable and tied to start time
sessionID can not be set via a config entry.  We still fallback to the
random number if it isn't set.  We also append the base 36 representation
of the session start time in seconds to the session ID.  This allows us
to identify a user across their sessions as well as within a session, which
helps when determining cache usage.