Set initial site_domain to null to avoid double cookies
On domains that have more than one dot in them, allowing the site_domain
to be set automatically will cause initial cookies to be set on the
wrong domain.  This will result in duplicate cookies (on different
subdomain levels) for the same site, and we'll always end up reading the
wrong cookie.

To get around this, we remove the default site_domain and wait for
config.js to load up before setting any cookie.