refresh cache based on boomerang version
The old way of refreshing cache, which used location.hash doesn't work
on Firefox 3.5.11 and breaks the back button on IE 8.

This new version refreshes based on the current and required versions of
boomerang.  We pass the version number in the query string so that this
file can be cached forever (at least for a given version).