Read and set cookie every time init is called
The first time init is called, site_domain may have been null, so we
wouldn't be able to set cookies.  The second time init is called,
initialized is true, so the setCookie call is never executed.  This
causes sessions to break.

We move the cookie checking and setting code up before the initialized
check.  This is all the code that can and should run on every call to
init... not just because it needs the latest data, but also because
config may change the cookie name it needs to read/set.