Set document.domain inside onload handler too
Setting document.domain in the iframe's src attribute only persists
until we call document.open() (IE <= 8).  This gives us enough access
to write JavaScript into the iframe, but not to execute it since
document.domain is reset at that time.

This patch sets document.domain again in the onload handler so that it
persists past the document.open().