Pass beacon URL to config.js
config.js needs to check if the server we will send the beacon
to is alive or not, and if not, redirect the beacon to a new
server.  To do this, we first need to send the beacon URL to
config.js.

This code does that, and makes sure we send the beacon URL from
the cookie and not the one hardcoded in boomerang.