Use our iframe's location instead of parent location
If you create an anonymous iframe, the browser sets location.href to
the value of the parent's location.href.  This is useful to get around
any permissions issues with trying to read the location of the parent
window.

This patch uses the boomerang iframe location.href for all its href
needs.