Fix document.domain issues with IE8
If we were loaded in an iframe, then we'd have a problem with IE8
and older, and possibly newer as well if the customer changed the
value of document.domain after the boomerang loader ran.

This patch adds a second check at the top of boomerang.  If this
check results in permission issues, then we try to fix the value
of document.domain.