Implement a clearMetrics method
This changeset adds a clearMetrics method that clears metrics after they've been
sent on the beacon.  This makes sure that the same metrics aren't sent to the
server more than once if multiple beacons fire on the same page.  This could
happen, for example with load & unload beacons, or xhr beacons.