Reset lastActionTime when we reset the session or on user action
A session may be reset within a page's lifetime (eg: for single-page apps),
and in these cases, the unload handler is never called, so lastActionTime
never gets updated.

Instead, the correct thing to do is to reset lastActionTime every time the
user interacts with the page as well as every time we reset the session (since
this is part of the reset).