Measure custom metrics & timers even if both boomerang and config.js load after onload
If both boomerang and config.js load after onload, we will end up calling init 3 times,
and done 2 times, but only the last call to init (after the last call to done) will have
the necessary configuration settings.

This changeset makes sure that custom metrics and timers are gathered in this event as well,
and that this plugin does not block the beacon from firing for other plugins.