Convert input from ISO-8859-1 to UTF-8
This is for Bug 82582

The SOASTA API returns data as ISO-8859-1, while the JSON parser expects UTF-8.
The problem we faced was that non-UTF-8 data was being ignored by the JSON parser,
so we ended up ignoring the original references.

This changeset converts the data using iconv before passing it to the JSON parser.