Move setting session params into sendBeacon
We've noticed issues where RT.done() was called before config.js
returned, and as a result an invalid session ID was set on the beacon.

This patch makes sure that the session ID is written to the beacon just
before we send the beacon.  This makes sure we have the correct ID on
the beacon.  We also set the session start time and length since these
are related.