Enable retries for UserTiming and ResourceTiming
Some resources or user timing marks/measures may be set after onload,
but before the beacon is fired.  This is especially true if the customer
writes a custom plugin that prevents the beacon from firing until they
are ready.

In these cases, we will need to check ResourceTiming and UserTiming
when boomerang is ready to send the beacon.

This would normally be done in the `before_beacon` event handler, but
we also need to make sure the Custom Timer code runs before
RT.addTimersToBeacon, which runs on the `befor_beacon` event.  Now,
since the page-params plugin is added before the rt plugin, we know
that PageParams' before_beacon handler should fire before RT's, however
this may not always be true.

To be safe, we run the retry() in the is_complete method, which is
called multiple times, but will always be called before `before_beacon`
fires.