Only look at edata if ename === "xhr"
Some websites hijack the browser's event model, and this results
in a broken event object coming through to our code.  Attempts to
access this event object throw exceptions.

This patch makes sure we only check the object when the event is
xhr, so we know that the data was generated by our own code.