Add a before_unload event that only fires onbeforeunload
We had problems where some code that caused beacons to be sent was running
both onbeforeunload as well as onunload.  Since the RT plugin is the only
one that does interesting things on these two events, it made sense to create
a new event called before_unload that only fires on the onbeforeunload event.

The old page_unload event remains the same and will fire on both beforeunload
as well as unload/pagehide

This patch also reverts all the hacks we put into place to prevent execution
on unload.