nodeCount can now return an object if required
If the second parameter to nodeCount is an array of keys, it will
return an object rather than just the value.  The first key will
map to the primary nodecount, and the second key, if any, will map
to the result of the nodecount after applying the passed in filter,
if any.

If no key array is passed in, behaviour is as before, if no filter
function is passed in, only a single element object will be returned
even if the key array contains two keys.