Workaround to send asynchronous beacons on unload as well.
We cannot send the beacon asynchronously on unload, so instead we check if we
are in the unload handler, and if we are, and if we have handled the last unload
handler, we will attempt to fire the beacon.

The beacon will only actually fire if one of the handlers queued it.  We still have
a problem if the last handler throws an exception.