Allow config overrides by the user
This set of changes adds the abillity for site admins to change
allowed and whitelisted changes in the configuration of Boomerang.

It hooks into BOOMR.init() and only overrides the configuration either
fetched from the server via config.js(on) or init() called by
BOOMR.plugins.logn_config.

With it comes a set of unit and E2E tests for cases where custom
metrics/pageParams are overwritten.

A whitelist of all allowed configuration changes is maintained in
`boomerang.js` under `impl.allowedConfigoverrides`

Necessairy for this change, there were 2 additions to the BOOMR.utils
set of utility functions, which are `utils.deepmerge` and
`filterByWhitelist`. Both have documentation inline.