Remove leading & trailing quotes from cookie
The Cookie spec states that cookies with embedded `=` or `&` should be quoted, and
some web servers (tomcat in particular) do not work well with unquoted cookies that
contain these characters.

This patch makes sure that any cookie that is quoted is unquoted once we pull it out
of the serialized cookie string.