Add support for XHR beacons being collected async
This fixes an error in th test case 18-autoxhr-before-page-load-alwayssendxhr.js
for SPAs where a second XHR would return faster than the first one thus
being collected and set as first index in the XHR beacon list. Since
tests in 18-autoxhr-before-page-load-alwayssendxhr.js used the array
index for testing availabillity of the request for a specific element
this was collected and tested incorrectly, which resulted in errors
during assertion.