Support updating beacon_url in domainAlias to primary domain Beacon URL (#406)
Since merely merging the session data from one domain to another may not completely suffice in the transferral of session information it was required to make sure that our sticky Beacon URL remains sticky during the transferral from primary to secondary domain. 

Tasked with this we also realized that for valid session statistics we are also required to merge over other data 
such as the summary load time over all session specific page loads and page loads that did not have a final page load timing (off by one errors).

This required more communication between plugins so that we were required to support updating the beacon URL and other params via the onconfig event in boomerang and distribute changes to the URL and RT data via the event system in Boomerang.