Rewrite mobile.js as a Boomerang plugin
The mobile.js plugin was not previously written as a plugin. This
commit updates the code to make it a plugin and implement the same
interface as other plugins. This allows for disabling the plugin as
needed and initializing the plugin in a consistent manner.

This commit also adds the `saveData` value from the Network Information
API if it is available.