Correctly detect our url on IE using non-standard loader
In all browsers except IE, document.currentScript will get us the current script node
In IE, if it uses our standard loader, one of the document.getElementById calls will get us the current script node
In IE 6-10, if not using the standard loader, we iterate through all script nodes in reverse, and stop at the first
whose readyState is "interactive", indicating that it is currently executing.
In IE 11, we throw an Error object, and pull the filename out of the error's stack property. This property is only
populated when an error is thrown.

Also add unit tests and page tests that test if BOOMR.url exists when using different kinds of loaders

Finally, updated comments that are no longer correct, and fix one instance where we used `document` rather
than `d` to check whether `visibilityState` was supported.